+++ docs/README.md
@@ -14,22 +14,26 @@ https://docs.python.org/3/library/unittest.html
 
 ## 例子
 
-* [单元测试](../test/test_unit.py)
+* [单元测试](../tests/test_unit.py)
 
-* [Selenium自动化测试](../test/test_selenium.py)
+* [Selenium自动化测试](../tests/test_selenium.py)
 
-* [requests接口测试](../test/test_api.py)
+* [requests接口测试](../tests/test_api.py)
 
-* [发送邮件](../test/test_send_email.py)
+* [发送邮件](../tests/test_send_email.py)
 
-* [发送钉钉消息](../test/test_send_dingtalk.py)
+* [发送钉钉消息](../tests/test_send_dingtalk.py)
 
-* [黑白名单](../test/test_label.py)
+* [黑白名单](../tests/test_label.py)
 
-* [XML格式的报告](../test/test_xml_report.py)
+* [XML格式的报告](../tests/test_xml_report.py)
 
 
 ## 使用文档
+
+* [HTML/XML测试报告](./test_report.md)
+  * [HTML测试报告](./test_report.md#HTML测试报告)
+  * [XML测试报告](./test_report.md#XML测试报告)
 
 * [不同类型的测试](./test_type.md)
   * [单元测试](./test_type.md#单元测试)
