--- test/test_unit.py
@@ -1,43 +1,0 @@
-import unittest
-from XTestRunner import HTMLTestRunner
-
-
-class TestDemo(unittest.TestCase):
-    """测试用例说明"""
-
-    def test_success(self):
-        """执行成功"""
-        self.assertEqual(2 + 3, 5)
-
-    @unittest.skip("skip case")
-    def test_skip(self):
-        pass
-
-    def test_fail(self):
-        self.assertEqual(5, 6)
-
-    def test_error(self):
-        self.assertEqual(a, 6)
-
-
-class TestDemo2(unittest.TestCase):
-
-    def test_success(self):
-        self.assertEqual(2 + 2, 4)
-
-
-class TestDemo3(unittest.TestCase):
-
-    def test_fail(self):
-        self.assertEqual(3, 4)
-
-
-if __name__ == '__main__':
-    report = "./reports/unit_result.html"
-    with(open(report, 'wb')) as fp:
-        unittest.main(testRunner=HTMLTestRunner(
-            stream=fp,
-            tester="虫师",
-            title='unittest单元测试测试报告',
-            description=['类型：单元测试']
-        ))
